apps: queue response body writes for better scheduling
Previously, as soon as a server received a request, it would try to emit
headers and body during the same cycle. When then body was too large to
send (e.g., due to congestion window) the response would be queued up
for later rounds of writing. This initial send always happened
regardless of a stream's selected priority.

The later rounds of writing explicitly use a writable iterator to step
through the streams. When switching to an iterator that better supports
priority scheduling, streams that are queued are more likely to respect
the selected priority.

With this change, we no longer attempt to write body immediately.
Instead, it is added to the disaptch queue that will be drained in
accordance with the writable iterator. Also, switch over to the
writable_prioritized iterator.