lib: progress through StreamIter from front to back
Previously, StreamIter next() would pop() the last stream ID and return
it, leading to a back-to-front progression.

Since we're stuck with SmallVec for internal data for performance reasons
we can't pop_front(). So instead, this change skips mutating the data
and instead uses a simple index that lets us move front-to-back.

This change helps to avoid cases where we had a correctly ordered
SmallVec of streams, and needed to reverse it before constructing a
StreamIter, jsut so that the iteration would make logical progressions.